# backend/static/openapi.yml
components:
  schemas:
    Author:
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        scholar_id:
          type: string
          nullable: true
        papers:
          type: array
          readOnly: true
          items:
            type: object
            properties:
              id:
                type: integer
                readOnly: true
              latest_citations:
                type: integer
                readOnly: true
                nullable: true
      type: object
    Citation:
      properties:
        id:
          type: integer
          readOnly: true
        num_cited:
          type: integer
        paper_id:
          type: integer
          readOnly: true
        date:
          type: string
          format: date-time
      type: object
    Paper:
      properties:
        id:
          type: integer
          readOnly: true
        year:
          type: integer
        name:
          type: string
        latest_citation:
          readOnly: true
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Citation'

      type: object
    Tag:
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
      type: object
...
    AuthorsTags:
      properties:
        authors:
          type: array
          items:
            type: integer
        tags:
          type: array
          items:
            type: integer
...
info:
  description: API Docs
  title: API Docs
  version: 1.0.0
openapi: 3.0.2
paths:
  /api/authors:
    post:
      tags:
        - Authors
      description: Create a new author
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                type: integer
    get:
      tags:
        - Authors
      description: Get all authors
      parameters:
        - in: query
          name: include
          required: false
          example: papers,tags
          schema:
            type: string
        - in: query
          name: name
          required: false
          schema:
            type: string
        - in: query
          name: tags
          required: false
          example: 1,2,3
          schema:
            type: string
        - in: query
          name: min-i10
          required: false
          schema:
            type: integer
        - in: query
          name: max-i10
          required: false
          schema:
            type: integer
        - in: query
          name: min-h
          required: false
          schema:
            type: integer
        - in: query
          name: max-h
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
  /api/authors/{author_id}:
    get:
      tags:
        - Authors
      description: Get an author
      parameters:
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Author not found
    put:
      tags:
        - Authors
      description: Update an author
      parameters:
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        '200':
          description: Updated successfully
        '404':
          description: Author not found
    delete:
      tags:
        - Authors
      description: Delete an author
      parameters:
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
        '404':
          description: Author not found
  /api/authors/{author_id}/papers:
    get:
      tags:
        - Authors
      description: Get all papers by author
      parameters:
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Paper'
        '404':
          description: Author not found
  /api/authors/{author_id}/papers/{paper_id}:
    put:
      tags:
        - Authors
      description: Add a paper to an author
      parameters:
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Added successfully
        '404':
          description: Author or paper not found
    delete:
      tags:
        - Authors
      description: Remove a paper from an author
      parameters:
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
        - in: path
          name: paper_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removed successfully
        '404':
          description: Author or paper not found
  /api/authors/{author_id}/tags:
    get:
      tags:
        - Authors
      description: Get all tags by author
      parameters:
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '404':
          description: Author not found
  /api/authors/{author_id}/tags/{tag_id}:
    put:
      tags:
        - Authors
      description: Add a tag to an author
      parameters:
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
        - in: path
          name: tag_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Added successfully
        '404':
          description: Author or tag not found
    delete:
      tags:
        - Authors
      description: Remove a tag from an author
      parameters:
        - in: path
          name: author_id
          required: true
          schema:
            type: integer
        - in: path
          name: tag_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removed successfully
        '404':
          description: Author or tag not found
  /api/authors/tags:
    put:
      tags:
        - Authors
      description: Batch add tags to authors
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorsTags'
      responses:
        '200':
          description: Tags added to authors
    delete:
      tags:
        - Authors
      description: Batch remove tags from authors
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorsTags'
      responses:
        '200':
          description: Tags removed from authors
...